Bug 385738 - Shorten the classpath/modulepath if necessary.
Shorten the classpath and/or modulepath if it gets longer than current
operating system limits.
Depending on the java version, os and launch configuration, the
classpath argument will be replaced by an argument file, a
classpath-only jar or an env variable.
The modulepath is replaced by an argument file if necessary.
If classpath-only jar needs to be used to shorten the classpath, an
explicit error popup ask if user wants to use classpath-only jar.

Change-Id: Ibcc23b6fa5c5baa05d91aef0f03529fb8e598e08